requirements.txt
cudf==0.6.1.post1
cupy==12.2.0
h5py==3.10.0
matplotlib==3.1.2
numpy==1.24.4
pandas==2.0.3
pyro==3.16
PyYAML==6.0.1
scikit_learn==1.3.1
scipy==1.11.3
torch==2.0.1
tqdm==4.30.0
